feat(plot): separation of plots in plot module
Closes #20.
