+++ bff/__init__.py
@@ -3,10 +3,19 @@ import logging
 
 from ._version import get_versions
 
+# Import submodules.
+from . import plot
+
 from .fancy import (
-    cast_to_category_pd, concat_with_categories, get_peaks, idict,
-    mem_usage_pd, parse_date, plot_history, plot_predictions, plot_series,
-    plot_true_vs_pred, read_sql_by_chunks, sliding_window, value_2_list
+    cast_to_category_pd,
+    concat_with_categories,
+    get_peaks,
+    idict,
+    mem_usage_pd,
+    parse_date,
+    read_sql_by_chunks,
+    sliding_window,
+    value_2_list,
 )
 
 from .config import FancyConfig
@@ -19,10 +28,7 @@ __all__ = [
     'idict',
     'mem_usage_pd',
     'parse_date',
-    'plot_history',
-    'plot_predictions',
-    'plot_series',
-    'plot_true_vs_pred',
+    'plot',
     'read_sql_by_chunks',
     'sliding_window',
     'value_2_list',
@@ -30,8 +36,8 @@ __all__ = [
 ]
 
 # Logging configuration.
-FORMAT = '%(asctime)-15s %(message)s'
-logging.basicConfig(format=FORMAT)
+FORMAT = '%(asctime)s [%(levelname)-7s] %(name)s: %(message)s'
+logging.basicConfig(format=FORMAT, level=logging.INFO)
 
 __version__ = get_versions()['version']
 del get_versions
