+++ bff/plot/__init__.py
@@ -1,0 +1,16 @@
+"""Plot module of bff."""
+
+from .plot import (
+    plot_history,
+    plot_predictions,
+    plot_series,
+    plot_true_vs_pred,
+)
+
+# Public object of the module.
+__all__ = [
+    'plot_history',
+    'plot_predictions',
+    'plot_series',
+    'plot_true_vs_pred',
+]
