+++ setup.py
@@ -26,6 +26,7 @@ REQUIRES = [
     'matplotlib==3.1.1',
     'numpy==1.16.4',
     'pandas==0.25.0',
+    'python-dateutil==2.8.0',
     'pyyaml==5.1.1',
     'scipy==1.3.0',
     'typing==3.7.4'
